Print a warning if web.config is detected
When there's a Web.config included in the source code, the developer
probably wants to launch their app with Hosted Web Core. That likely
means they want to use the hwc-buildpack. However, there are legititate
cases to use the binary_buildpack with Web.config, so we only print this
message when there is no start command or Procfile.

Signed-off-by: Matthew Horan <mhoran@pivotal.io>